requirements.txt
numpy>=1.20.1
Pillow>=8.1.2
matplotlib>=3.3.4
opencv-python>=4.5.1.48